refactor(패키지): 패키지 세분화
- LottoApplilcationMain에 집중되어 있는 비즈니스 로직 분산 위해 패키지 분산
- domain 밑에 lotto, result 패키지 생성
