+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domian;
+package lotto.domain.lotto;
 
 import java.util.HashSet;
 import java.util.List;
