+++ src/main/java/lotto/domain/lotto/LottoBundle.java
@@ -1,4 +1,4 @@
-package lotto.domian;
+package lotto.domain.lotto;
 
 import java.util.List;
