+++ src/main/java/lotto/domain/lotto/LottoGroup.java
@@ -1,4 +1,4 @@
-package lotto.domian;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
