+++ src/main/java/lotto/domain/lotto/service/LottoService.java
@@ -1,0 +1,8 @@
+package lotto.domain.lotto.service;
+
+import lotto.domain.lotto.Money;
+
+public class LottoService {
+
+
+}
