+++ src/main/java/lotto/domain/result/ProfitRate.java
@@ -1,4 +1,4 @@
-package lotto.domian;
+package lotto.domain.result;
 
 public class ProfitRate {
