+++ src/main/java/lotto/domain/result/WinNumber.java
@@ -1,4 +1,7 @@
-package lotto.domian;
+package lotto.domain.result;
+
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 
 public class WinNumber {
