+++ src/test/java/lotto/domain/LottoCompanyTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domian.LottoCompany;
+import lotto.domain.lotto.LottoCompany;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
