+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domian.LottoNumber;
+import lotto.domain.lotto.LottoNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
