+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
-import lotto.domian.Lotto;
-import lotto.domian.LottoNumber;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
