+++ src/test/java/lotto/domain/MatchCountTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domian.MatchCount;
+import lotto.domain.result.MatchCount;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
