+++ src/test/java/lotto/domain/ProfitCalculatorTest.java
@@ -1,6 +1,9 @@
 package lotto.domain;
 
-import lotto.domian.*;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoBundle;
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.result.*;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
