+++ src/test/java/lotto/domain/RecordTest.java
@@ -1,6 +1,11 @@
 package lotto.domain;
 
-import lotto.domian.*;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoBundle;
+import lotto.domain.lotto.LottoCompany;
+import lotto.domain.result.Rank;
+import lotto.domain.result.Record;
+import lotto.domain.result.WinNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
